requirements.txt
pandas
numpy
sentence-transformers
fastapi
uvicorn
openai
pyyaml
